pyproject.toml
[project]
name = "trainer-bot"
version = "0.1.0"
description = "Self-hostable Telegram bot that acts as a personal training/health advisor backed by intervals.icu and Groq LLM."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Oleksii Lasiichuk" }]
dependencies = [
  "python-telegram-bot[rate-limiter]>=22.0",
  "openai>=1.50.0",
  "httpx>=0.27.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "sqlalchemy>=2.0",
  "greenlet>=3.0",
  "aiosqlite>=0.20",
  "structlog>=24.0",
  "tenacity>=9.0",
  "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
  "ruff>=0.5",
  "mypy>=1.10",
]
postgres = ["asyncpg>=0.29", "psycopg[binary]>=3.2"]

[project.scripts]
trainer-bot = "trainer_bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trainer_bot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
